+++ Source/Scene/Cesium3DTileset.js
@@ -63,6 +63,8 @@ import TileOrientedBoundingBox from "./TileOrientedBoundingBox.js";
  * @param {Resource|String|Promise<Resource>|Promise<String>} options.url The url to a tileset JSON file.
  * @param {Boolean} [options.show=true] Determines if the tileset will be shown.
  * @param {Matrix4} [options.modelMatrix=Matrix4.IDENTITY] A 4x4 transformation matrix that transforms the tileset's root tile.
+ * @param {Axis} [options.modelUpAxis=Axis.Y] Which axis is considered up when loading models for tile contents.
+ * @param {Axis} [options.modelForwardAxis=Axis.X] Which axis is considered forward when loading models for tile contents.
  * @param {ShadowMode} [options.shadows=ShadowMode.ENABLED] Determines whether the tileset casts or receives shadows from light sources.
  * @param {Number} [options.maximumScreenSpaceError=16] The maximum screen space error used to drive level of detail refinement.
  * @param {Number} [options.maximumMemoryUsage=512] The maximum amount of memory in MB that can be used by the tileset.
@@ -102,6 +104,7 @@ import TileOrientedBoundingBox from "./TileOrientedBoundingBox.js";
  * @param {String|Number} [options.instanceFeatureIdLabel="instanceFeatureId_0"] Label of the instance feature ID set used for picking and styling. If instanceFeatureIdLabel is set to an integer N, it is converted to the string "instanceFeatureId_N" automatically. If both per-primitive and per-instance feature IDs are present, the instance feature IDs take priority.
  * @param {Boolean} [options.showCreditsOnScreen=false] Whether to display the credits of this tileset on screen.
  * @param {SplitDirection} [options.splitDirection=SplitDirection.NONE] The {@link SplitDirection} split to apply to this tileset.
+ * @param {Boolean} [options.projectTo2D=false] Whether to accurately project the tileset to 2D. If this is true, the tileset will be projected accurately to 2D, but it will use more memory to do so. If this is false, the tileset will use less memory and will still render in 2D / CV mode, but its projected positions may be inaccurate. This cannot be set after the tileset has loaded.
  * @param {String} [options.debugHeatmapTilePropertyName] The tile variable to colorize as a heatmap. All rendered tiles will be colorized relative to each other's specified variable value.
  * @param {Boolean} [options.debugFreezeFrame=false] For debugging only. Determines if only the tiles from last frame should be used for rendering.
  * @param {Boolean} [options.debugColorizeTiles=false] For debugging only. When true, assigns a random color to each tile.
@@ -163,7 +166,8 @@ function Cesium3DTileset(options) {
   this._geometricError = undefined; // Geometric error when the tree is not rendered at all
   this._extensionsUsed = undefined;
   this._extensions = undefined;
-  this._gltfUpAxis = undefined;
+  this._modelUpAxis = undefined;
+  this._modelForwardAxis = undefined;
   this._cache = new Cesium3DTilesetCache();
   this._processingQueue = [];
   this._selectedTiles = [];
@@ -481,7 +485,7 @@ function Cesium3DTileset(options) {
    *         return;
    *     }
    *
-   *     console.log('Loading: requests: ' + numberOfPendingRequests + ', processing: ' + numberOfTilesProcessing);
+   *     console.log(`Loading: requests: ${numberOfPendingRequests}, processing: ${numberOfTilesProcessing}`);
    * });
    */
   this.loadProgress = new Event();
@@ -589,8 +593,8 @@ function Cesium3DTileset(options) {
    *
    * @example
    * tileset.tileFailed.addEventListener(function(error) {
-   *     console.log('An error occurred loading tile: ' + error.url);
-   *     console.log('Error: ' + error.message);
+   *     console.log(`An error occurred loading tile: ${error.url}`);
+   *     console.log(`Error: ${error.message}`);
    * });
    */
   this.tileFailed = new Event();
@@ -769,6 +773,8 @@ function Cesium3DTileset(options) {
     options.splitDirection,
     SplitDirection.NONE
   );
+
+  this._projectTo2D = defaultValue(options.projectTo2D, false);
 
   /**
    * This property is for debugging only; it is not optimized for production use.
@@ -998,16 +1004,22 @@ function Cesium3DTileset(options) {
       }
 
       that._root = that.loadTileset(resource, tilesetJson);
+
+      // Handle legacy gltfUpAxis option
       const gltfUpAxis = defined(tilesetJson.asset.gltfUpAxis)
         ? Axis.fromName(tilesetJson.asset.gltfUpAxis)
         : Axis.Y;
+      const modelUpAxis = defaultValue(options.modelUpAxis, gltfUpAxis);
+      const modelForwardAxis = defaultValue(options.modelForwardAxis, Axis.X);
+
       const asset = tilesetJson.asset;
       that._asset = asset;
       that._properties = tilesetJson.properties;
       that._geometricError = tilesetJson.geometricError;
       that._extensionsUsed = tilesetJson.extensionsUsed;
       that._extensions = tilesetJson.extensions;
-      that._gltfUpAxis = gltfUpAxis;
+      that._modelUpAxis = modelUpAxis;
+      that._modelForwardAxis = modelForwardAxis;
       that._extras = tilesetJson.extras;
 
       const extras = asset.extras;
@@ -1155,8 +1167,8 @@ Object.defineProperties(Cesium3DTileset.prototype, {
    * @exception {DeveloperError} The tileset is not loaded.  Use Cesium3DTileset.readyPromise or wait for Cesium3DTileset.ready to be true.
    *
    * @example
-   * console.log('Maximum building height: ' + tileset.properties.height.maximum);
-   * console.log('Minimum building height: ' + tileset.properties.height.minimum);
+   * console.log(`Maximum building height: ${tileset.properties.height.maximum}`);
+   * console.log(`Minimum building height: ${tileset.properties.height.minimum}`);
    *
    * @see Cesium3DTileFeature#getProperty
    * @see Cesium3DTileFeature#setProperty
@@ -1610,8 +1622,7 @@ Object.defineProperties(Cesium3DTileset.prototype, {
 
   /**
    * The total amount of GPU memory in bytes used by the tileset. This value is estimated from
-   * geometry, texture, and batch table textures of loaded tiles. For point clouds, this value also
-   * includes per-point metadata.
+   * geometry, texture, batch table textures, and binary metadata of loaded tiles.
    *
    * @memberof Cesium3DTileset.prototype
    *