+++ Source/Scene/GltfLoader.js
@@ -78,7 +78,7 @@ const GltfLoaderState = {
  * @param {Axis} [options.upAxis=Axis.Y] The up-axis of the glTF model.
  * @param {Axis} [options.forwardAxis=Axis.Z] The forward-axis of the glTF model.
  * @param {Boolean} [options.loadAttributesAsTypedArray=false] Load all attributes and indices as typed arrays instead of GPU buffers.
- * @param {Boolean} [options.loadPositionsFor2D=false] If true, load the positions buffer as a typed array for accurately projecting models to 2D.
+ * @param {Boolean} [options.loadAttributesFor2D=false] If true, load the positions buffer and any instanced attribute buffers as typed arrays for accurately projecting models to 2D.
  * @param {Boolean} [options.loadIndicesForWireframe=false] If true, load the index buffer as both a buffer and typed array. The latter is useful for creating wireframe indices in WebGL1.
  * @param {Boolean} [options.renameBatchIdSemantic=false] If true, rename _BATCHID or BATCHID to _FEATURE_ID_0. This is used for .b3dm models
  * @private
@@ -100,7 +100,7 @@ export default function GltfLoader(options) {
     options.loadAttributesAsTypedArray,
     false
   );
-  const loadPositionsFor2D = defaultValue(options.loadPositionsFor2D, false);
+  const loadAttributesFor2D = defaultValue(options.loadAttributesFor2D, false);
   const loadIndicesForWireframe = defaultValue(
     options.loadIndicesForWireframe,
     false
@@ -126,7 +126,7 @@ export default function GltfLoader(options) {
   this._upAxis = upAxis;
   this._forwardAxis = forwardAxis;
   this._loadAttributesAsTypedArray = loadAttributesAsTypedArray;
-  this._loadPositionsFor2D = loadPositionsFor2D;
+  this._loadAttributesFor2D = loadAttributesFor2D;
   this._loadIndicesForWireframe = loadIndicesForWireframe;
   this._renameBatchIdSemantic = renameBatchIdSemantic;
 
@@ -405,14 +405,11 @@ function loadVertexBuffer(
   semantic,
   draco,
   dequantize,
-  loadAsTypedArray,
-  loadFor2D
+  loadBuffer,
+  loadTypedArray
 ) {
   const accessor = gltf.accessors[accessorId];
   const bufferViewId = accessor.bufferView;
-
-  const loadBuffer = !loadAsTypedArray;
-  const loadTypedArray = loadAsTypedArray || loadFor2D;
 
   const vertexBufferLoader = ResourceCache.loadVertexBuffer({
     gltf: gltf,
@@ -640,21 +637,13 @@ function getSetIndex(gltfSemantic) {
   return undefined;
 }
 
-function loadAttribute(
-  loader,
-  gltf,
-  accessorId,
-  semanticType,
-  gltfSemantic,
-  draco,
-  dequantize,
-  loadAsTypedArray,
-  loadAsTypedArrayPacked,
-  frameState
-) {
-  const accessor = gltf.accessors[accessorId];
-  const bufferViewId = accessor.bufferView;
+const scratchSemanticInfo = {
+  gltfSemantic: undefined,
+  renamedSemantic: undefined,
+  modelSemantic: undefined,
+};
 
+function getSemanticInfo(loader, semanticType, gltfSemantic) {
   // For .b3dm, rename _BATCHID (or the legacy BATCHID) to _FEATURE_ID_0
   // in the generated model components for compatibility with EXT_mesh_features
   let renamedSemantic = gltfSemantic;
@@ -665,11 +654,39 @@ function loadAttribute(
     renamedSemantic = "_FEATURE_ID_0";
   }
 
-  const name = gltfSemantic;
   const modelSemantic = semanticType.fromGltfSemantic(renamedSemantic);
+
+  const semanticInfo = scratchSemanticInfo;
+  semanticInfo.gltfSemantic = gltfSemantic;
+  semanticInfo.renamedSemantic = renamedSemantic;
+  semanticInfo.modelSemantic = modelSemantic;
+
+  return semanticInfo;
+}
+
+function loadAttribute(
+  loader,
+  gltf,
+  accessorId,
+  semanticInfo,
+  draco,
+  dequantize,
+  loadBuffer,
+  loadTypedArray,
+  loadAsTypedArrayPacked
+) {
+  const accessor = gltf.accessors[accessorId];
+  const bufferViewId = accessor.bufferView;
+
+  const gltfSemantic = semanticInfo.gltfSemantic;
+  const renamedSemantic = semanticInfo.renamedSemantic;
+  const modelSemantic = semanticInfo.modelSemantic;
+
   const setIndex = defined(modelSemantic)
     ? getSetIndex(renamedSemantic)
     : undefined;
+
+  const name = gltfSemantic;
   const attribute = createAttribute(
     gltf,
     accessorId,
@@ -682,11 +699,6 @@ function loadAttribute(
     return attribute;
   }
 
-  const loadFor2D =
-    modelSemantic === VertexAttributeSemantic.POSITION &&
-    loader._loadPositionsFor2D &&
-    !frameState.scene3DOnly;
-
   const vertexBufferLoader = loadVertexBuffer(
     loader,
     gltf,
@@ -694,8 +706,8 @@ function loadAttribute(
     gltfSemantic,
     draco,
     dequantize,
-    loadAsTypedArray,
-    loadFor2D
+    loadBuffer,
+    loadTypedArray
   );
   const promise = vertexBufferLoader.promise.then(function (
     vertexBufferLoader
@@ -715,6 +727,10 @@ function loadAttribute(
       );
       attribute.byteOffset = 0;
       attribute.byteStride = undefined;
+
+      if (loadBuffer) {
+        attribute.buffer = vertexBufferLoader.buffer;
+      }
     } else {
       attribute.typedArray = vertexBufferLoader.typedArray;
       attribute.buffer = vertexBufferLoader.buffer;
@@ -744,19 +760,39 @@ function loadVertexAttribute(
   accessorId,
   gltfSemantic,
   draco,
+  hasInstances,
   frameState
 ) {
+  const semanticInfo = getSemanticInfo(
+    loader,
+    VertexAttributeSemantic,
+    gltfSemantic
+  );
+
+  const modelSemantic = semanticInfo.modelSemantic;
+  const isPositionAttribute =
+    modelSemantic === VertexAttributeSemantic.POSITION;
+  const loadFor2D =
+    isPositionAttribute &&
+    !hasInstances &&
+    loader._loadAttributesFor2D &&
+    !frameState.scene3DOnly;
+
+  const loadAsTypedArrayOnly = loader._loadAttributesAsTypedArray;
+  const loadBuffer = !loadAsTypedArrayOnly;
+  const loadTypedArray = loadAsTypedArrayOnly || loadFor2D;
+  const loadAsTypedArrayPacked = false;
+
   return loadAttribute(
     loader,
     gltf,
     accessorId,
-    VertexAttributeSemantic,
-    gltfSemantic,
+    semanticInfo,
     draco,
     false,
-    loader._loadAttributesAsTypedArray,
-    false,
-    frameState
+    loadBuffer,
+    loadTypedArray,
+    loadAsTypedArrayPacked
   );
 }
 
@@ -765,17 +801,37 @@ function loadInstancedAttribute(
   gltf,
   accessorId,
   gltfSemantic,
-  loadAsTypedArrayPacked
+  loadAsTypedArrayPacked,
+  frameState
 ) {
+  const semanticInfo = getSemanticInfo(
+    loader,
+    InstanceAttributeSemantic,
+    gltfSemantic
+  );
+
+  const modelSemantic = semanticInfo.modelSemantic;
+
+  const isTranslationAttribute =
+    modelSemantic === InstanceAttributeSemantic.TRANSLATION;
+
+  const loadFor2D =
+    isTranslationAttribute &&
+    loader._loadAttributesFor2D &&
+    !frameState.scene3DOnly;
+
+  const loadBuffer = !loadAsTypedArrayPacked;
+  loadAsTypedArrayPacked = loadAsTypedArrayPacked || loadFor2D;
+
   // Don't pass in draco object since instanced attributes can't be draco compressed
   return loadAttribute(
     loader,
     gltf,
     accessorId,
-    InstanceAttributeSemantic,
-    gltfSemantic,
+    semanticInfo,
     undefined,
     true,
+    loadBuffer,
     loadAsTypedArrayPacked,
     loadAsTypedArrayPacked
   );
@@ -1122,6 +1178,7 @@ function loadPrimitive(
   loader,
   gltf,
   gltfPrimitive,
+  hasInstances,
   supportedImageFormats,
   frameState
 ) {
@@ -1155,6 +1212,7 @@ function loadPrimitive(
             accessorId,
             semantic,
             draco,
+            hasInstances,
             frameState
           )
         );
@@ -1380,7 +1438,7 @@ function loadInstances(loader, gltf, nodeExtensions, frameState) {
       defined(gltf.accessors[attributes.TRANSLATION].max);
     for (const semantic in attributes) {
       if (attributes.hasOwnProperty(semantic)) {
-        // Load the attributes as typed arrays if:
+        // In addition to the loader options, load the attributes as typed arrays if:
         // - the instances have rotations, so that instance matrices are computed on the CPU.
         //   This avoids the expensive quaternion -> rotation matrix conversion in the shader.
         // - the translation accessor does not have a min and max, so the values can be used
@@ -1405,7 +1463,8 @@ function loadInstances(loader, gltf, nodeExtensions, frameState) {
             gltf,
             accessorId,
             semantic,
-            loadAsTypedArrayPacked
+            loadAsTypedArrayPacked,
+            frameState
           )
         );
       }
@@ -1506,6 +1565,16 @@ function loadNode(loader, gltf, gltfNode, supportedImageFormats, frameState) {
   node.rotation = fromArray(Quaternion, gltfNode.rotation);
   node.scale = fromArray(Cartesian3, gltfNode.scale);
 
+  const nodeExtensions = defaultValue(
+    gltfNode.extensions,
+    defaultValue.EMPTY_OBJECT
+  );
+  const instancingExtension = nodeExtensions.EXT_mesh_gpu_instancing;
+
+  if (defined(instancingExtension)) {
+    node.instances = loadInstances(loader, gltf, nodeExtensions, frameState);
+  }
+
   const meshId = gltfNode.mesh;
   if (defined(meshId)) {
     const mesh = gltf.meshes[meshId];
@@ -1517,6 +1586,7 @@ function loadNode(loader, gltf, gltfNode, supportedImageFormats, frameState) {
           loader,
           gltf,
           primitives[i],
+          defined(node.instances),
           supportedImageFormats,
           frameState
         )
@@ -1534,16 +1604,6 @@ function loadNode(loader, gltf, gltfNode, supportedImageFormats, frameState) {
     node.morphWeights = defined(morphWeights)
       ? morphWeights.slice()
       : arrayFill(new Array(targetsLength), 0.0);
-  }
-
-  const nodeExtensions = defaultValue(
-    gltfNode.extensions,
-    defaultValue.EMPTY_OBJECT
-  );
-  const instancingExtension = nodeExtensions.EXT_mesh_gpu_instancing;
-
-  if (defined(instancingExtension)) {
-    node.instances = loadInstances(loader, gltf, nodeExtensions, frameState);
   }
 
   return node;