+++ Specs/Scene/ModelExperimental/WireframePipelineStageSpec.js
@@ -1,6 +1,7 @@
 import {
   combine,
   GltfLoader,
+  ModelExperimentalStatistics,
   PrimitiveType,
   Resource,
   ResourceCache,
@@ -77,6 +78,7 @@ describe(
         wireframeIndexBuffer: undefined,
         model: {
           _resources: resources,
+          statistics: new ModelExperimentalStatistics(),
         },
       };
     }
@@ -100,7 +102,14 @@ describe(
         expect(renderResources.primitiveType).toBe(PrimitiveType.TRIANGLES);
 
         WireframePipelineStage.process(renderResources, primitive, frameState);
-        expect(renderResources.wireframeIndexBuffer).toBeDefined();
+
+        const wireframeIndexBuffer = renderResources.wireframeIndexBuffer;
+        const model = renderResources.model;
+        expect(wireframeIndexBuffer).toBeDefined();
+        expect(model._resources).toEqual([wireframeIndexBuffer]);
+        expect(model.statistics.geometryByteLength).toBe(
+          wireframeIndexBuffer.sizeInBytes
+        );
         expect(renderResources.primitiveType).toBe(PrimitiveType.LINES);
         expect(renderResources.count).toBe(72);
       });
@@ -124,7 +133,15 @@ describe(
         expect(renderResources.primitiveType).toBe(PrimitiveType.TRIANGLES);
 
         WireframePipelineStage.process(renderResources, primitive, frameState);
-        expect(renderResources.wireframeIndexBuffer).toBeDefined();
+
+        const wireframeIndexBuffer = renderResources.wireframeIndexBuffer;
+        const model = renderResources.model;
+        expect(wireframeIndexBuffer).toBeDefined();
+        expect(model._resources).toEqual([wireframeIndexBuffer]);
+        expect(model.statistics.geometryByteLength).toBe(
+          wireframeIndexBuffer.sizeInBytes
+        );
+
         expect(renderResources.primitiveType).toBe(PrimitiveType.LINES);
         expect(renderResources.count).toBe(72);
       });